Removing routes to multi part form
