+++ src/App.tsx
@@ -11,12 +11,8 @@ import SchemeWoundUp from './views/SchemeWoundUp';
 import SchemeStatusAndMembership from './views/SchemeStatusAndMembership';
 import ElectronicComsConsent from './views/ElectronicComsConsent';
 import SchemeNameAndAddress from './views/SchemeNameAndAddress';
-import IndividualTrusteeStepOne from './views/AddIndividualTrustee/IndividualTrusteeStepOne';
 import Trustees from './views/Trustees';
 import AddTrustee from './views/AddTrustee';
-import IndividualTrusteeStepTwo from './views/AddIndividualTrustee/IndividualTrusteeStepTwo';
-import IndividualTrusteeStepThree from './views/AddIndividualTrustee/IndividualTrusteeStepThree';
-import IndividualTrusteeStepFour from './views/AddIndividualTrustee/IndividualTrusteeStepFour';
 
 function App() {
   return (
@@ -76,26 +72,6 @@ function App() {
               exact
               path="/trustees/add-trustee"
               component={() => <AddTrustee />}
-            />
-            <Route
-              exact
-              path="/trustees/add-trustee/add-individual-step-1"
-              component={() => <IndividualTrusteeStepOne />}
-            />
-            <Route
-              exact
-              path="/trustees/add-trustee/add-individual-step-2"
-              component={() => <IndividualTrusteeStepTwo />}
-            />
-            <Route
-              exact
-              path="/trustees/add-trustee/add-individual-step-3"
-              component={() => <IndividualTrusteeStepThree />}
-            />
-            <Route
-              exact
-              path="/trustees/add-trustee/add-individual-step-4"
-              component={() => <IndividualTrusteeStepFour />}
             />
           </Switch>
         </AppWidth>
