+++ src/views/AddIndividualTrustee/IndividualTrusteeStepFour.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import Styles from '../Layout.module.scss';
-import UserResearchSidebar from '../../components/UserResearchSidebar';
 import { ArrowLink, ArrowButton } from '@tpr/layout';
 import { useHistory } from 'react-router-dom';
 import { H3, H1, Hr } from '@tpr/core';
@@ -41,36 +40,33 @@ const IndividualTrusteeStepFour = () => {
     },
   ];
   return (
-    <div className={Styles.root}>
-      <UserResearchSidebar />
-      <div className={Styles.main}>
-        <ArrowLink
-          pointsTo="left"
-          iconSide="left"
-          title="Back"
-          onClick={() => history.goBack()}
-          cfg={{ mt: 3 }}
-        />
-        <H3 cfg={{ mt: 2 }}>Add trustee: Individual</H3>
-        <H1 cfg={{ mt: 2, mb: 3 }}>Contact details for the trustee</H1>
-        <Hr cfg={{ mt: 4, mb: 5 }} />
-        <Form onSubmit={submit}>
-          {({ handleSubmit, valid }) => (
-            <form onSubmit={handleSubmit}>
-              {renderFields(fields)}
-              <Hr cfg={{ mt: 8, mb: 4 }} />
-              <ArrowButton
-                title="Save and Continue"
-                type="submit"
-                pointsTo="right"
-                iconSide="right"
-                disabled={!valid}
-              />
-            </form>
-          )}
-        </Form>
-        <ScrollToTop />
-      </div>
+    <div className={Styles.main}>
+      <ArrowLink
+        pointsTo="left"
+        iconSide="left"
+        title="Back"
+        onClick={() => history.goBack()}
+        cfg={{ mt: 3 }}
+      />
+      <H3 cfg={{ mt: 2 }}>Add trustee: Individual</H3>
+      <H1 cfg={{ mt: 2, mb: 3 }}>Contact details for the trustee</H1>
+      <Hr cfg={{ mt: 4, mb: 5 }} />
+      <Form onSubmit={submit}>
+        {({ handleSubmit, valid }) => (
+          <form onSubmit={handleSubmit}>
+            {renderFields(fields)}
+            <Hr cfg={{ mt: 8, mb: 4 }} />
+            <ArrowButton
+              title="Save and Continue"
+              type="submit"
+              pointsTo="right"
+              iconSide="right"
+              disabled={!valid}
+            />
+          </form>
+        )}
+      </Form>
+      <ScrollToTop />
     </div>
   );
 };
