+++ src/views/AddIndividualTrustee/IndividualTrusteeStepThree.tsx
@@ -1,18 +1,13 @@
 import React from 'react';
 import Styles from '../Layout.module.scss';
-import UserResearchSidebar from '../../components/UserResearchSidebar';
 import { ArrowLink, ArrowButton } from '@tpr/layout';
 import { useHistory } from 'react-router-dom';
 import { H3, H1, Hr } from '@tpr/core';
 import { Form, renderFields, FieldProps } from '@tpr/forms';
 import ScrollToTop from '../../components/ScrollToTop';
 
-const IndividualTrusteeStepThree = () => {
+const IndividualTrusteeStepThree = (props: any) => {
   const history = useHistory();
-
-  const submit = () => {
-    history.push('add-individual-step-4');
-  };
 
   const fields: FieldProps[] = [
     {
@@ -71,37 +66,34 @@ const IndividualTrusteeStepThree = () => {
     },
   ];
   return (
-    <div className={Styles.root}>
-      <UserResearchSidebar />
-      <div className={Styles.main}>
-        <ArrowLink
-          pointsTo="left"
-          iconSide="left"
-          title="Back"
-          onClick={() => history.goBack()}
-          cfg={{ mt: 3 }}
-        />
-        <H3 cfg={{ mt: 2 }}>Add trustee: Individual</H3>
-        <H1 cfg={{ mt: 2, mb: 3 }}>What is this trustee's address?</H1>
-        <Hr cfg={{ mt: 4, mb: 5 }} />
-        <Form onSubmit={submit}>
-          {({ handleSubmit, valid }) => (
-            <form onSubmit={handleSubmit}>
-              {renderFields(fields)}
-              <Hr cfg={{ mt: 8, mb: 4 }} />
-              <ArrowButton
-                type="submit"
-                iconSide="right"
-                pointsTo="right"
-                title="Continue"
-                cfg={{ my: 4 }}
-                disabled={!valid}
-              />
-            </form>
-          )}
-        </Form>
-        <ScrollToTop />
-      </div>
+    <div className={Styles.main}>
+      <ArrowLink
+        pointsTo="left"
+        iconSide="left"
+        title="Back"
+        onClick={() => history.goBack()}
+        cfg={{ mt: 3 }}
+      />
+      <H3 cfg={{ mt: 2 }}>Add trustee: Individual</H3>
+      <H1 cfg={{ mt: 2, mb: 3 }}>What is this trustee's address?</H1>
+      <Hr cfg={{ mt: 4, mb: 5 }} />
+      <Form onSubmit={props.onSubmit}>
+        {({ handleSubmit, valid }) => (
+          <form onSubmit={handleSubmit}>
+            {renderFields(fields)}
+            <Hr cfg={{ mt: 8, mb: 4 }} />
+            <ArrowButton
+              type="submit"
+              iconSide="right"
+              pointsTo="right"
+              title="Continue"
+              cfg={{ my: 4 }}
+              disabled={!valid}
+            />
+          </form>
+        )}
+      </Form>
+      <ScrollToTop />
     </div>
   );
 };
