+++ src/views/AddIndividualTrustee/IndividualTrusteeStepTwo.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import Styles from '../Layout.module.scss';
-import UserResearchSidebar from '../../components/UserResearchSidebar';
 import { ArrowLink, ArrowButton } from '@tpr/layout';
 import { useHistory } from 'react-router-dom';
 import { H1, H3, Hr, H4 } from '@tpr/core';
@@ -13,12 +12,8 @@ import {
 } from '@tpr/forms';
 import ScrollToTop from '../../components/ScrollToTop';
 
-const IndividualTrusteeStepTwo = () => {
+const IndividualTrusteeStepTwo = (props: any) => {
   const history = useHistory();
-
-  const submit = (values: any) => {
-    history.push('add-individual-step-3');
-  };
 
   const TRUSTEE_TYPE_BUTTONS = 'trusteeType';
   const PROF_TRUSTEE_BUTTONS = 'professionalTrustee';
@@ -56,49 +51,42 @@ const IndividualTrusteeStepTwo = () => {
     ],
   ];
   return (
-    <div className={Styles.root}>
-      <UserResearchSidebar />
-      <div className={Styles.main}>
-        <ArrowLink
-          pointTo="left"
-          iconSide="left"
-          title="Back"
-          cfg={{ mt: 3 }}
-          onClick={() => history.goBack()}
-        />
-        <H3 cfg={{ mt: 2 }}>Add trustee: Individual</H3>
-        <H1 cfg={{ mt: 2, mb: 3 }}>Type of trustee</H1>
-        <Hr cfg={{ mt: 4, mb: 5 }} />
-        <Form onSubmit={submit}>
-          {({ handleSubmit, values }) => (
-            <form onSubmit={handleSubmit}>
-              <H4 cfg={{ mb: 2 }}>
-                Select the option that best describes the type of trustee:
-              </H4>
-              <SeparatorY>{renderFields(fields[0])}</SeparatorY>
-              <br />
-              <H4 cfg={{ mb: 2 }}>
-                Is this individual a professional trustee?
-              </H4>
-              <SeparatorX>{renderFields(fields[1])}</SeparatorX>
-              <Hr cfg={{ mt: 9, mb: 5 }} />
-              <ArrowButton
-                title="Continue"
-                pointsTo="right"
-                iconSide="right"
-                type="submit"
-                cfg={{ my: 4 }}
-                disabled={
-                  !(
-                    values[TRUSTEE_TYPE_BUTTONS] && values[PROF_TRUSTEE_BUTTONS]
-                  )
-                }
-              />
-            </form>
-          )}
-        </Form>
-        <ScrollToTop />
-      </div>
+    <div className={Styles.main}>
+      <ArrowLink
+        pointTo="left"
+        iconSide="left"
+        title="Back"
+        cfg={{ mt: 3 }}
+        onClick={() => history.goBack()}
+      />
+      <H3 cfg={{ mt: 2 }}>Add trustee: Individual</H3>
+      <H1 cfg={{ mt: 2, mb: 3 }}>Type of trustee</H1>
+      <Hr cfg={{ mt: 4, mb: 5 }} />
+      <Form onSubmit={props.onSubmit}>
+        {({ handleSubmit, values }) => (
+          <form onSubmit={handleSubmit}>
+            <H4 cfg={{ mb: 2 }}>
+              Select the option that best describes the type of trustee:
+            </H4>
+            <SeparatorY>{renderFields(fields[0])}</SeparatorY>
+            <br />
+            <H4 cfg={{ mb: 2 }}>Is this individual a professional trustee?</H4>
+            <SeparatorX>{renderFields(fields[1])}</SeparatorX>
+            <Hr cfg={{ mt: 9, mb: 5 }} />
+            <ArrowButton
+              title="Continue"
+              pointsTo="right"
+              iconSide="right"
+              type="submit"
+              cfg={{ my: 4 }}
+              disabled={
+                !(values[TRUSTEE_TYPE_BUTTONS] && values[PROF_TRUSTEE_BUTTONS])
+              }
+            />
+          </form>
+        )}
+      </Form>
+      <ScrollToTop />
     </div>
   );
 };
